pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-decode"
version = "1.0.0"
description = "A tool for extracting structured information from multilingual documents using Google's Gemini Pro Vision API"
readme = "Readme.md"
authors = [{name = "GeminiDecode Team", email = "example@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.28.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.0.300",
    "PyPDF2>=3.0.1",
    "chromadb>=0.4.0",
    "faiss-cpu>=1.7.4",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/geminidecode"
Repository = "https://github.com/yourusername/geminidecode"
"Bug Tracker" = "https://github.com/yourusername/geminidecode/issues"

[project.scripts]
gemini-decode = "app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gemini_decode*"]

[tool.setuptools.package-data]
"*" = ["*.md", "LICENSE", ".streamlit/*"]